Conflicts:
	weasis-acquire/weasis-acquire-explorer/src/main/java/org/weasis/acquire/explorer/gui/dialog/AcquirePublishDialog.java
	weasis-core/weasis-core-api/src/main/java/org/weasis/core/api/image/util/ImageFiler.java
	weasis-core/weasis-core-ui/src/main/java/org/weasis/core/ui/editor/image/DefaultView2d.java
	weasis-dicom/weasis-dicom-codec/src/main/java/org/weasis/dicom/codec/display/ShutterOp.java

+++ weasis-acquire/weasis-acquire-explorer/src/main/java/org/weasis/acquire/explorer/util/ImageInfoHelper.java
@@ -45,7 +45,7 @@ public class ImageInfoHelper {
             ImageElement imgElt = imgInfo.getImage();
             Integer width = (Integer) imgElt.getTagValue(TagW.ImageWidth);
             Integer height = (Integer) imgElt.getTagValue(TagW.ImageHeight);
-            Double currentImageSize = (double) Math.max(width, height);
+            double currentImageSize = Math.max(width, height);
             return BigDecimal.valueOf(expectedImageSize / currentImageSize).setScale(5, RoundingMode.HALF_UP)
                 .doubleValue();
         } catch (NullPointerException e) {
